Move messages to root intl file
